Implement the get{,s}_len commands.
This change implements one new command family, get{,s}_len.  This is similar
to the get command with the added bonus of being able to specify to the
server how many bytes of the VALUE you want sent back if the KEY is found.
These new commands are very useful when the structure of the data you are
caching is highly structured and repetitive (i.e., a list of packed 64-bit
unmarshaled values).